requirements.txt
scikit-image
streamlit
numpy
opencv-python
Pillow